+++ src/main/java/com/overclock/overclock/service/OverclockService.java
@@ -2,10 +2,12 @@ package com.overclock.overclock.service;
 
 import com.overclock.overclock.model.Overclock;
 
+import java.math.BigInteger;
+
 public interface OverclockService {
-    Overclock getOverclockById(int id);
+    Overclock getOverclockById(BigInteger id);
     boolean save(Overclock overclock);
-    boolean update(int id, Overclock newOverclock);
-    Overclock getDefaultValues(int overclockId, int assemblyId);
+    boolean update(BigInteger id, Overclock newOverclock);
+    Overclock getDefaultValues(BigInteger overclockId, BigInteger assemblyId);
     boolean validate(Overclock overclock);
 }
